default props and applay eslint rules
