+++ package.json
@@ -39,6 +39,7 @@
     "lodash": "^4.17.4",
     "prop-types": "^15.6.0",
     "react": "^16.0.0",
+    "react-autosuggest": "^9.3.2",
     "react-dom": "^16.0.0"
   },
   "devDependencies": {
@@ -69,7 +70,6 @@
     "pre-commit": "^1.2.2",
     "raf": "^3.4.0",
     "react-addons-test-utils": "^15.6.2",
-    "react-autosuggest": "^9.3.2",
     "react-proptypes": "^1.0.0",
     "react-test-renderer": "^15.4.1",
     "style-loader": "^0.13.1",
