+++ src/constants/defaultProps.js
@@ -1,3 +1,16 @@
 export const fieldDefaultProps = {
-
+    callbacks: {},
+    className: '',
+    errorStyles: {
+        className: '',
+        itemClassName: ''
+    },
+    fieldAttributes: {},
+    hasValidationError: false,
+    label: '',
+    name: '',
+    placeholder: '',
+    validationErrors: [],
+    value: '',
+    wrapperClassName: '',
 };
