+++ tests/unit/helpers.test.js
@@ -72,149 +72,4 @@ describe('helpers', () => {
             expect(clonedObject).toEqual(objectWithArrays);
         });
     });
-
-    describe('isNotEqualObject', () => {
-        it('should return false when one of objects is null', () => {
-            const srcObject = null;
-            const compareObject = { foo: 'bar', foo2: 'bar2' };
-            expect(isNotEqualObject(srcObject, compareObject)).toBeFalsy();
-        });
-        it('should return true when objects has diffrent length', () => {
-            const srcObject = { foo: 'bar' };
-            const compareObject = { foo: 'bar', foo2: 'bar2' };
-            expect(isNotEqualObject(srcObject, compareObject)).toBeTruthy();
-        });
-        it('should return true when objects keys has diffrent types', () => {
-            const srcObject = { foo: 'bar' };
-            const compareObject = { foo: true };
-            expect(isNotEqualObject(srcObject, compareObject)).toBeTruthy();
-        });
-        it('should return true when objects keys has diffrent values', () => {
-            const srcObject = { foo: 'bar' };
-            const compareObject = { foo: 'bar2' };
-            expect(isNotEqualObject(srcObject, compareObject)).toBeTruthy();
-        });
-        it('should return true when objects keys are objects and has diffrent values', () => {
-            const srcObject = {
-                foo: {
-                    foo2: 'bar',
-                },
-            };
-            const compareObject = {
-                foo: {
-                    foo2: 'bar2',
-                },
-            };
-            expect(isNotEqualObject(srcObject, compareObject)).toBeTruthy();
-        });
-        it('should return true when objects keys are arrays and has diffrent lengths', () => {
-            const srcObject = {
-                foo: [
-                    'bar',
-                    'bar2',
-                ],
-            };
-            const compareObject = {
-                foo: [
-                    'bar',
-                ],
-            };
-            expect(isNotEqualObject(srcObject, compareObject)).toBeTruthy();
-        });
-        it('should return true when objects keys are arrays and has diffrent values', () => {
-            const srcObject = {
-                foo: [
-                    'bar',
-                    'bar2',
-                ],
-            };
-            const compareObject = {
-                foo: [
-                    'bar',
-                    'bar3',
-                ],
-            };
-            expect(isNotEqualObject(srcObject, compareObject)).toBeTruthy();
-        });
-
-
-        it('should return true when objects keys has no diffrent values', () => {
-            const srcObject = { foo: 'bar' };
-            const compareObject = { foo: 'bar' };
-            expect(isNotEqualObject(srcObject, compareObject)).toBeFalsy();
-        });
-        it('should return true when object properties are objects and has the same values', () => {
-            const srcObject = {
-                foo: {
-                    foo2: 'bar',
-                },
-            };
-            const compareObject = {
-                foo: {
-                    foo2: 'bar',
-                },
-            };
-            expect(isNotEqualObject(srcObject, compareObject)).toBeFalsy();
-        });
-        it('should return true when objects properties are arrays and has the same lengths and values', () => {
-            const srcObject = {
-                foo: [
-                    'bar',
-                ],
-            };
-            const compareObject = {
-                foo: [
-                    'bar',
-                ],
-            };
-            expect(isNotEqualObject(srcObject, compareObject)).toBeFalsy();
-        });
-        it('should return true when objects properties are the same components', () => {
-            const TestComponent = () => (<span>test</span>);
-            expect (isNotEqualObject(TestComponent, TestComponent)).toBeFalsy();
-        });
-    });
-
-    describe('isReactComponentOrElement', () => {
-        class ClassComponent extends React.Component {
-            render() {
-                return (
-                    <span>test</span>
-                );
-            }
-        }
-        const StatlessComponent = () => (
-            <span>test</span>
-        );
-
-        const nodeElement = <span>test</span>;
-
-        it('should return true if detect class component', () => {
-            expect(isReactComponentOrElement(ClassComponent)).toEqual(true);
-        });
-
-        it('should return true if detect stateless component', () => {
-            expect(isReactComponentOrElement(StatlessComponent)).toEqual(true);
-        });
-
-        it('should return true if detect node element', () => {
-            expect(isReactComponentOrElement(nodeElement)).toEqual(true);
-        });
-
-        it('should return false if detect object that is not react element', () => {
-            expect(isReactComponentOrElement({})).toEqual(false);
-        });
-
-        it('should return false if detect arrow function that is not react component', () => {
-            expect(isReactComponentOrElement(() => {})).toEqual(false);
-        });
-
-        it('should return false if detect function that is not react component', () => {
-            expect(isReactComponentOrElement(jest.fn())).toEqual(false);
-        });
-
-        it('should return false if detect string that is not react component', () => {
-            expect(isReactComponentOrElement('foo')).toEqual(false);
-        });
-    });
 });
